retrieve: move results file to pending only if results file is not already in pending

+++ scripts/retrieve_domains.sh
@@ -141,11 +141,8 @@ filter() {
 # source. The output is a cumulative filtered domains file of all filtered
 # domains from all sources in this run.
 process_source() {
-    if [[ "$results_file" != data/pending/* ]]; then
-        # Move results file to pending directory
-        mv "$results_file" data/pending
-        results_file="data/pending/${results_file}"
-    fi
+    [[ -f "$results_file" ]] && mv "$results_file" data/pending
+    results_file="data/pending/${results_file}"
 
     [[ ! -f "$results_file" ]] && return
